Added project structure/modularisation
