+++ src-tauri/src/commands/mod.rs
@@ -1,0 +1,2 @@
+pub mod save_video;
+pub mod set_wallpaper;
